add methods in taskServiceInterface and implement them

+++ src/main/java/com/api/service/TaskServiceInterface.java
@@ -46,6 +46,6 @@ public interface TaskServiceInterface {
      * @param taskId of the task
      * @return optional of a task
      */
-    Optional<TaskWithLogDto> findByTaskIdWithLog(Long taskId);
+    Optional<TaskWithLogDto> findByTaskIdWithLog(Integer taskId);
 
 }
